requirements.txt
django ==3.0.4
djangorestframework ==3.11.0
django-cors-headers ==3.2.0 
channels == 2.4.0
channels-redis ==2.4.0
redis ==2.4.0
